Struct thread_scoped::JoinGuard

#[must_use = "thread will be immediately joined if `JoinGuard` is not used"]
pub struct JoinGuard<'a, T: Send + 'a> {
    // some fields omitted
}

A RAII guard for that joins a scoped thread upon drop

Panics

JoinGuard will panic on join or drop if its owned thread panics

Methods

impl<'a, T: Send + 'a> JoinGuard<'a, T>

fn thread(&self) -> &Thread

Provides the backing Thread object

fn join(self) -> T

Joins the guarded thread and returns its result

Panics

join() will panic if the owned thread panics